requirements.txt
setuptools==77.0.3
amqp==5.3.1
celery==5.4.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.9
pika==1.3.2
pytest==8.3.5
pytest-django==4.10.0
python-decouple==3.8
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==5.5.0
djoser==2.3.1
drf_extra_fields==3.7.0
minio==7.2.15
pyyaml==6.0.2
psycopg2-binary==2.9.10
django-timescaledb==0.2.13
django-celery-beat==2.7.0
pyfakefs==5.7.4
python-magic==0.4.27
psutil==7.0.0
mock==5.2.0
grpcio==1.70.0
grpcio-tools==1.70.0
daphne==4.1.2
channels==4.2.0
starlette==0.46.0